Request page from Drupal on 404 error
This change makes it so that when the React application hits a 404 status code error it will request the url from the Drupal application and render the content into the react application's body.

It also parses the meta tags from the response and inserts them into the head of the application using next.js Head component.